// src/api/generated/models/dashboardOrderStatusDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { DashboardOrderStatusDtoStatus } from './dashboardOrderStatusDtoStatus';

export interface DashboardOrderStatusDto {
  status: DashboardOrderStatusDtoStatus;
  count: number;
}